# frontend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Keel CRM API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/pipelines/{id}/stages/": {
            "post": {
                "operationId": "pipelines_stages_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "pipelines"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pipeline"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/pipelines/{id}/stages/reorder/": {
            "post": {
                "operationId": "pipelines_stages_reorder_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "pipelines"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pipeline"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pipeline": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "position": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "is_default": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "stages": {
                        "readOnly": true
                    },
                    "version": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "archived_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "archived_at",
                    "created_at",
                    "id",
                    "is_default",
                    "name",
                    "position",
                    "stages",
                    "updated_at",
                    "version"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "keel_session"
            }
        }
    }
}